pyproject.toml
[tool.poetry]
name = "llm-multi-ai-agent-system"
version = "0.1.0"
description = "LLM multi-agent system pipeline for stock price prediction using AWS services"
authors = ["Long Hoang <longhhoang1202@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
# Core AWS and config
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
# Logging (following copilot instructions)
loguru = "^0.7.0"
# LLM framework with GPT-5-mini support (following copilot instructions)
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
openai = "^1.30.0"
# Use newer tiktoken version with better wheel support
tiktoken = "^0.8.0"
# Data processing
pandas = "^2.0.0"
numpy = "^1.24.0"
# HTTP client
aiohttp = "^3.9.0"
# Validation
pydantic = "^2.0.0"
# Financial data
alpha-vantage = "^2.3.1"
# ML
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.23.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
# AWS mocking for tests (following copilot instructions)
moto = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true